Move `PlottingThreadPriority` to `shared` module
